# tools/swagger/docs/api-internal/iam-internal/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Ces APIS sont utilisés pour configurer les clients, tenants, idp, groupes, profils, et utilisateurs.",
    "version": "v1",
    "title": "API IAM - Gestion des accès aux identités",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Direction de la diffusion et des partenariats",
      "email": "contact@programmevitam.fr"
    },
    "license": {
      "name": "License to be defined ...",
      "url": "http://notedefined.com"
    }
  },
  "host": "localhost",
  "basePath": "/iam/v1",
...
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/customers": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Récupération paginée des clients",
        "operationId": "getAllPaginatedUsingGET",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "$ref": "#/parameters/criteria"
          },
          {
            "$ref": "#/parameters/orderBy"
          },
          {
            "$ref": "#/parameters/direction"
          },
          {
            "name": "embedded",
            "in": "query",
            "description": "Non utilisées",
            "required": false,
            "type": "string",
            "enum": [
              "ALL"
            ]
          },
          {
            "$ref": "#/parameters/page"
          },
          {
            "$ref": "#/parameters/size"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PaginatedValuesDto«CustomerDto»"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Création d'un client",
        "operationId": "createUsingPOST",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "in": "body",
            "name": "customer",
            "description": "Client à créer",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CustomerDto"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/CustomerDto"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          }
        }
      }
    },
...
    "/customers/{id}/history": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Récupération de l'historique d'un client",
        "operationId": "findHistoryByIdUsingGET",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant technique du client",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonNode"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
    "/groups": {
      "get": {
        "tags": [
          "Groupes"
        ],
        "summary": "Récupération paginée des groupes",
        "operationId": "getAllPaginatedUsingGET_1",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "$ref": "#/parameters/criteria"
          },
          {
            "$ref": "#/parameters/orderBy"
          },
          {
            "$ref": "#/parameters/direction"
          },
          {
            "name": "embedded",
            "in": "query",
            "description": "Récupération des profiles associés",
            "required": false,
            "type": "string",
            "enum": [
              "ALL"
            ]
          },
          {
            "$ref": "#/parameters/page"
          },
          {
            "$ref": "#/parameters/size"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PaginatedValuesDto«GroupDto»"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "Groupes"
        ],
        "summary": "Création d'un groupe",
        "operationId": "createUsingPOST_1",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "in": "body",
            "name": "groupe",
            "description": "Groupe à créer",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GroupDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GroupDto"
            }
          },
          "401": {
            "description": "La requête n'est pas autorisée. Le X-User-Token n'est pas valide"
          },
          "403": {
            "description": "L'utilisateur ne possède pas les droits pour effectuer cette opération"
          }
        }
      }
    },
...
    "/groups/levels": {
      "get": {
        "tags": [
          "Groupes"
        ],
        "summary": "Récupération des niveaux des groupes",
        "operationId": "getLevelsUsingGET",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "$ref": "#/parameters/criteria"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
...
    "/groups/{id}/history": {
      "get": {
        "tags": [
          "Groupes"
        ],
        "summary": "Récupération de l'historique d'un groupe",
        "operationId": "findHistoryByIdUsingGET_1",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant technique du groupe",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonNode"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
...
    "/owners/{id}/history": {
      "get": {
        "tags": [
          "Propriétaires"
        ],
        "summary": "Récupération de l'historique d'un propriétaire",
        "operationId": "findHistoryByIdUsingGET_2",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant technique du propriétaire",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonNode"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
    "/profiles": {
      "get": {
        "tags": [
          "Profils"
        ],
        "summary": "Récupération paginée des profils",
        "operationId": "getAllPaginatedUsingGET_2",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "$ref": "#/parameters/criteria"
          },
          {
            "$ref": "#/parameters/orderBy"
          },
          {
            "$ref": "#/parameters/direction"
          },
          {
            "name": "embedded",
            "in": "query",
            "description": "Récupération des compteurs usersCount, groupCount ainsi que le nom du tenant",
            "required": false,
            "type": "string",
            "enum": [
              "ALL"
            ]
          },
          {
            "$ref": "#/parameters/page"
          },
          {
            "$ref": "#/parameters/size"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PaginatedValuesDto«ProfileDto»"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "Profils"
        ],
        "summary": "Création d'un profil",
        "operationId": "createUsingPOST_4",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "in": "body",
            "name": "profile",
            "description": "Un profile",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProfileDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProfileDto"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          }
        }
      }
    },
...
    "/profiles/levels": {
      "get": {
        "tags": [
          "Profils"
        ],
        "summary": "Récupération des niveaux des profils",
        "operationId": "getLevelsUsingGET_1",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "$ref": "#/parameters/criteria"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
...
    "/profiles/{id}/history": {
      "get": {
        "tags": [
          "Profils"
        ],
        "summary": "Récupération de l'historique d'un profil",
        "operationId": "findHistoryByIdUsingGET_3",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant technique du profil",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonNode"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "Utilisateurs"
        ],
        "summary": "Récupération paginée des utilisateurs",
        "operationId": "getAllPaginatedUsingGET_3",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "$ref": "#/parameters/page"
          },
          {
            "$ref": "#/parameters/size"
          },
          {
            "$ref": "#/parameters/criteria"
          },
          {
            "$ref": "#/parameters/orderBy"
          },
          {
            "$ref": "#/parameters/direction"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste paginée des utilisateurs",
            "schema": {
              "$ref": "#/definitions/PaginatedValuesDto«UserDto»"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "Utilisateurs"
        ],
        "summary": "Création d'un utilisateur",
        "operationId": "createUsingPOST_7",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "in": "body",
            "name": "user",
            "description": "Utilisateur à créer",
            "required": false,
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "L'utilisateur a été créé avec succès",
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          }
        }
      }
    },
...
    "/users/levels": {
      "get": {
        "tags": [
          "Utilisateurs"
        ],
        "summary": "Récupération des niveaux des utilisateurs",
        "operationId": "getLevelsUsingGET_2",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "in": "body",
            "name": "criteria",
            "description": "Critères de recherche",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des niveaux des utilisateurs",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
...
    "/users/{id}/history": {
      "get": {
        "tags": [
          "Utilisateurs"
        ],
        "summary": "Récupération de l'historique d'un utilisateur",
        "operationId": "findHistoryByIdUsingGET_4",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant technique de l'utilisateur",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Historique d'un utilisateur",
            "schema": {
              "$ref": "#/definitions/JsonNode"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    }
  },
  "responses": {
    "Unauthorized": {
      "description": "La requête n'est pas autorisée. Le X-User-Token n'est pas valide"
    },
    "NotFound": {
      "description": "Ressource non trouvée"
    }
  },
  "parameters": {
    "page": {
      "in": "query",
      "name": "page",
      "required": true,
      "minimum": 0,
      "type": "integer",
      "format": "int32",
      "description": "Numéro de la page"
    },
    "size": {
      "name": "size",
      "in": "query",
      "required": true,
      "minimum": 1,
      "type": "integer",
      "format": "int32",
      "description": "Nombre de résultats maximum"
    },
    "criteria": {
      "name": "criteria",
      "in": "query",
      "description": "Critères de recherche",
      "required": false,
      "type": "string"
    },
    "orderBy": {
      "name": "orderBy",
      "in": "query",
      "description": "Nom du champ sur lequel effectuer le tri",
      "required": false,
      "type": "string"
    },
    "direction": {
      "name": "direction",
      "in": "query",
      "description": "Direction pour le tri",
      "required": false,
      "type": "string",
      "enum": [
        "ASC",
        "DESC"
      ]
    },
    "x-tenant-id": {
      "name": "X-Tenant-Id",
      "in": "header",
      "description": "Identifiant du tenant sur lequel doit être exécuté la requête",
      "required": true,
      "type": "integer",
      "format": "int32"
    },
    "x-user-token": {
      "name": "X-User-Token",
      "in": "header",
      "description": "Identifiant technique du token généré par le serveur CAS lors de l'authentification",
      "required": true,
      "type": "string"
    },
    "x-application-id": {
      "name": "X-Application-Id",
      "in": "header",
      "description": "Identifiant unique utilisé dans la génération des logs et des journaux VITAM",
      "required": false,
      "type": "string"
    },
    "x-user-level": {
      "name": "X-User-Level",
      "in": "header",
      "description": "Niveau de l'utilisateur",
      "required": true,
      "type": "string"
    },
    "x-customer-id": {
      "name": "X-Customer-Id",
      "in": "header",
      "description": "Identifiant technique du client",
      "required": true,
      "type": "string"
    }
  },
  "definitions": {
...
    "CustomerDto": {
      "type": "object",
      "properties": {
        "address": {
          "$ref": "#/definitions/AddressDto"
        },
        "code": {
          "type": "string",
          "example": "1",
          "description": "Identifiant métier"
        },
        "companyName": {
          "type": "string",
          "example": "archiveco",
          "description": "Nom de la compagnie"
        },
        "defaultEmailDomain": {
          "type": "string",
          "example": "@archiveco.fr",
          "description": "Nom de domaine par défaut"
        },
        "emailDomains": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "Ensemble des domaines",
          "example": [
            "@archiveo.fr",
            "@archiveco.com"
          ]
        },
        "enabled": {
          "type": "boolean",
          "example": "true",
          "description": "Actif ou inactif"
        },
        "id": {
          "type": "string",
          "description": "Identifiant technique",
          "example": "1",
          "readOnly": true
        },
        "identifier": {
          "type": "string",
          "example": "1",
          "description": "Identifiant métier",
          "readOnly": true
        },
        "language": {
          "type": "string",
          "enum": [
            "FRENCH",
            "ENGLISH",
            "GERMANY"
          ],
          "example": "FRENCH",
          "description": "Préférence de langue pour les utilisateurs du client"
        },
        "name": {
          "type": "string",
          "example": "archiveco",
          "description": "Nom du client"
        },
        "otp": {
          "type": "string",
          "enum": [
            "OPTIONAL",
            "DISABLED",
            "MANDATORY"
          ],
          "example": "OPTIONAL",
          "description": "Authentification en 2 étapes pour les utilisateurs du client"
        },
        "owners": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OwnerDto"
          },
          "description": "Propriétaires du client",
          "readOnly": true
        },
        "readonly": {
          "type": "boolean",
          "description": "Lecture seule",
          "example": true,
          "readOnly": true
        },
        "passwordRevocationDelay": {
          "type": "integer",
          "format": "int32",
          "example": "1",
          "description": "Durée de validitée des mot de passes pour les utilisateurs clients (exprimée en jours)"
        },
        "subrogeable": {
          "type": "boolean",
          "example": true,
          "description": "Utilisateur du client subrogeable ou non"
        },
        "hasCustomGraphicIdentity": {
          "type": "boolean"
        }
      },
      "required": [
        "code",
        "name",
        "companyName",
        "language",
        "passwordRevocationDelay",
        "otp",
        "emailDomains",
        "defaultEmailDomain",
        "address",
        "hasCustomGraphicIdentity"
      ],
      "title": "CustomerDto"
    },
...
    "JsonNode": {
      "type": "object",
      "properties": {
        "array": {
          "type": "boolean"
        },
        "bigDecimal": {
          "type": "boolean"
        },
        "bigInteger": {
          "type": "boolean"
        },
        "binary": {
          "type": "boolean"
        },
        "boolean": {
          "type": "boolean"
        },
        "containerNode": {
          "type": "boolean"
        },
        "double": {
          "type": "boolean"
        },
        "float": {
          "type": "boolean"
        },
        "floatingPointNumber": {
          "type": "boolean"
        },
        "int": {
          "type": "boolean"
        },
        "integralNumber": {
          "type": "boolean"
        },
        "long": {
          "type": "boolean"
        },
        "missingNode": {
          "type": "boolean"
        },
        "nodeType": {
          "type": "string",
          "enum": [
            "ARRAY",
            "BINARY",
            "BOOLEAN",
            "MISSING",
            "NULL",
            "NUMBER",
            "OBJECT",
            "POJO",
            "STRING"
          ]
        },
        "null": {
          "type": "boolean"
        },
        "number": {
          "type": "boolean"
        },
        "object": {
          "type": "boolean"
        },
        "pojo": {
          "type": "boolean"
        },
        "short": {
          "type": "boolean"
        },
        "textual": {
          "type": "boolean"
        },
        "valueNode": {
          "type": "boolean"
        }
      },
      "title": "JsonNode"
    },
...
    "PaginatedValuesDto«CustomerDto»": {
      "type": "object",
      "properties": {
        "hasMore": {
          "type": "boolean"
        },
        "pageNum": {
          "type": "integer",
          "format": "int32"
        },
        "pageSize": {
          "type": "integer",
          "format": "int32"
        },
        "values": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CustomerDto"
          }
        }
      },
      "title": "PaginatedValuesDto«CustomerDto»"
    },
...
  }
}